+++ src/lib.rs
@@ -13,6 +13,106 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! This crate aims to provide [Sigstore](https://www.sigstore.dev/) capabilities to Rust developers.
+//!
+//! Currently, the main focus of the crate is to provide the verification
+//! capabilities offered by the official `cosign` tool.
+//!
+//! **Warning:** this library is still experimental. Its API can change at any time.
+//!
+//! # Security
+//!
+//! Should you discover any security issues, please refer to
+//! Sigstore's [security process](https://github.com/sigstore/community/blob/main/SECURITY.md).
+//!
+//! # Verification
+//!
+//! Sigstore verification is done using the [`cosign::Client`](crate::cosign::Client)
+//! struct.
+//!
+//! ## Triangulation of Sigstore signature
+//!
+//! Given a container image/oci artifact, calculate the location of
+//! its cosign signature inside of a registry:
+//!
+//! ```rust,no_run
+//! use crate::sigstore::cosign::CosignCapabilities;
+//! use std::fs;
+//!
+//! #[tokio::main]
+//! pub async fn main() {
+//!   let auth = &sigstore::registry::Auth::Anonymous;
+//!
+//!   let mut client = sigstore::cosign::ClientBuilder::default()
+//!     .build()
+//!     .expect("Unexpected failure while building Client");
+//!   let image = "registry-testing.svc.lan/kubewarden/disallow-service-nodeport:v0.1.0";
+//!   let (cosign_signature_image, source_image_digest) = client.triangulate(
+//!     image,
+//!     auth
+//!   ).await.expect("Unexpected failure while using triangulate");
+//! }
+//! ```
+//!
+//! ## Signature verification
+//!
+//! Verify the signature of a container image/oci artifact:
+//!
+//! ```rust,no_run
+//! use crate::sigstore::cosign::CosignCapabilities;
+//! use std::collections::HashMap;
+//! use std::fs;
+//!
+//! #[tokio::main]
+//! pub async fn main() {
+//!   let auth = &sigstore::registry::Auth::Anonymous;
+//!
+//!   // Provide both rekor and fulcio data -> this enables keyless verification
+//!   // Read rekor's key from the location generated by `cosign initialize`
+//!   let rekor_pub_key = fs::read_to_string("~/.sigstore/root/targets/rekor.pub")
+//!     .expect("Cannot read rekor public key");
+//!   // Read fulcio's certificate from the location generated by `cosign initialize`
+//!   let fulcio_cert = fs::read_to_string("~/.sigstore/root/targets/fulcio.crt.pem")
+//!     .expect("Cannot read fulcio certificate");
+//!
+//!   let mut client = sigstore::cosign::ClientBuilder::default()
+//!     .with_rekor_pub_key(&rekor_pub_key)
+//!     .with_fulcio_cert(fulcio_cert.as_bytes())
+//!     .build()
+//!     .expect("Unexpected failure while building Client");
+//!
+//!   // Obtained via `triangulate`
+//!   let cosign_image = "registry-testing.svc.lan/kubewarden/disallow-service-nodeport:sha256-5f481572d088dc4023afb35fced9530ced3d9b03bf7299c6f492163cb9f0452e.sig";
+//!   // Obtained via `triangulate`
+//!   let source_image_digest = "sha256-5f481572d088dc4023afb35fced9530ced3d9b03bf7299c6f492163cb9f0452e";
+//!
+//!   let mut annotations: HashMap<String, String> = HashMap::new();
+//!   annotations.insert("env".to_string(), "prod".to_string());
+//!
+//!   let verification_key = fs::read_to_string("~/cosign.pub")
+//!     .expect("Cannot read contents of cosign public key");
+//!   let signatures_matching_requirements = client.verify(
+//!     auth,
+//!     cosign_image,
+//!     source_image_digest,
+//!     &Some(verification_key),
+//!     Some(annotations)
+//!   ).await.expect("unexpected verification error");
+//!
+//!   if signatures_matching_requirements.is_empty() {
+//!     panic!("no signature is matching the requirements");
+//!   } else {
+//!     println!("signatures matching the requirements: {:?}",
+//!         signatures_matching_requirements);
+//!   }
+//! }
+//! ```
+//!
+//! # Examples
+//!
+//! Additional examples can be found inside of the [`examples`](https://github.com/sigstore/sigstore-rs/tree/main/examples/)
+//! directory.
+
 mod crypto;
 mod mock_client;
